requirements.txt
torch==2.2.2+cpu
faiss-cpu==1.7.4
sentence-transformers==2.2.2
llama-cpp-python==0.2.11
numpy==1.26.4
scikit-learn==1.3.2
transformers==4.37.2
pandas==2.2.1
Flask==3.0.3
Jinja2==3.1.3
requests==2.31.0
tqdm==4.66.2
protobuf==4.25.3
typing_extensions==4.11.0
streamlit==1.32.2
